requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pytz
pyngrok
requests